refactor(grading): modularize grading logic
